Add simple can example
This is a simple example of a CAN node using the CSP library.

It includes two nodes: node1 and node2.

Node1 sends a message every second, and node2 receives it.

Node2 sends a reply to the node 1 as an acknowledgment.

The example uses the CAN protocol to communicate
between the two nodes.

The example is designed to be run on a Linux
system with the SocketCAN interface.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>